cast real8 data to double (not float) Original: e5aedf71d5b5b3920ce105da9511cfed28adb07d

+++ lalapps/src/lalapps/lalappsfrutils.c
@@ -78,7 +78,7 @@ FrameH *fr_add_proc_REAL8TimeSeries (
     fdata.step = (double) chan->deltaT;
     fdata.unit = unit;
     fdata.size = (size_t) chan->data->length;
-    fdata.ddata = (float *) chan->data->data;
+    fdata.ddata = (double *) chan->data->data;
     return fr_add_proc_data( frame, &fdata );
 }
